cod_bo6_analyzer.py
import cv2
import torch
import tensorflow as tf  # For TensorFlow Lite
import numpy as np
import mediapipe as mp
from google.cloud import storage
from google.cloud import language_v1
from google.cloud import videointelligence_v1


class CODAnalyzer:
    def __init__(self, video_path, yolov7_model_path="yolov7.pt", tflite_model_path="yolov7_quant.tflite"):  # Updated model paths

        self.cap = cv2.VideoCapture(video_path)

        # Load the YOLOv7 model
        self.yolov7_model = torch.hub.load('ultralytics/yolov5', 'yolov7', pretrained=True, _verbose=False)

        # Load the TensorFlow Lite model
        self.tflite_model = tf.lite.Interpreter(model_path=tflite_model_path)
        self.tflite_model.allocate_tensors()
        self.tflite_input_details = self.tflite_model.get_input_details()
        self.tflite_output_details = self.tflite_model.get_output_details()

        # ... (Initialize other models, e.g., for pose estimation or action recognition) ...

        # ... (Initialize connections to Google Cloud services and the Gemini API) ...

    def preprocess(self, frame):
        """
        Preprocesses the frame by resizing and converting to RGB.
        """
        resized_frame = cv2.resize(frame, (640, 640))
        rgb_frame = cv2.cvtColor(resized_frame, cv2.COLOR_BGR2RGB)
        return rgb_frame

    def detect_objects_yolov7(self, frame):
        """
        Performs object detection using the YOLOv7 model.
        """
        results = self.yolov7_model(frame)
        detections = results.pandas().xyxy[0]
        return detections

    def detect_objects_tflite(self, frame):
        """
        Performs object detection using the TensorFlow Lite model.
        """
        input_data = np.expand_dims(frame, axis=0)
        input_data = input_data.astype(np.float32)

        self.tflite_model.set_tensor(self.tflite_input_details[0]['index'], input_data)
        self.tflite_model.invoke()

        output_data = self.tflite_model.get_tensor(self.tflite_output_details[0]['index'])

        # Post-process the output data
        results = self.yolov7_model(frame)  # Use the regular YOLOv7 model for post-processing
        detections = results.pandas().xyxy[0]  # Use the same DataFrame structure as YOLOv7
        detections['confidence'] = output_data[..., 4]  # Replace confidence scores with TFLite output
        detections['class'] = output_data[..., 5].astype(int)  # Replace class IDs with TFLite output

        return detections

    def track_objects(self, detections, previous_frame):
        """
        Tracks objects across frames using StrongSORT.
        """
        # ... (Implement StrongSORT tracking here) ...
        return []  # Return an empty list for now

    def analyze_actions(self, tracked_objects, current_frame, previous_frame):
        """
        Analyzes player actions using pose estimation, etc.
        """
        # ... (Implement action recognition logic here) ...
        return None  # Return None for now

    def analyze_minimap(self, frame):
        """
        Extracts and analyzes information from the minimap.
        """
        # ... (Implement minimap analysis logic here) ...
        return None  # Return None for now

    def analyze_with_gemini(self, data):
        """
        Sends data to the Gemini API for analysis (NLP, code generation).
        """
        # ... (Implement Gemini API interaction here) ...
        return None  # Return None for now

    def draw_visualizations(self, frame, detections, tracked_objects, actions, minimap_data, gemini_results):
        """
        Draws bounding boxes, labels, and other visualizations on the frame.
        """
        # Example: Draw bounding boxes and labels from YOLOv7 detections
        for i in range(len(detections)):
            xmin, ymin, xmax, ymax, confidence, class_id, label = detections.iloc[i]

            if confidence > 0.5:  # Filter detections by confidence
                cv2.rectangle(frame, (int(xmin), int(ymin)), (int(xmax), int(ymax)), (0, 255, 0), 2)
                cv2.putText(frame, label, (int(xmin), int(ymin) - 10), cv2.FONT_HERSHEY_SIMPLEX, 0.9, (0, 255, 0), 2)
        return frame

    def run(self):
        """
        Runs the main analysis loop.
        """
        previous_frame = None  # Initialize previous frame for tracking

        while self.cap.isOpened():
            ret, frame = self.cap.read()
            if not ret:
                break

            processed_frame = self.preprocess(frame)

            # Choose appropriate object detection method
            if use_tflite:  # Make sure to define 'use_tflite' somewhere
                detections = self.detect_objects_tflite(processed_frame)
            else:
                detections = self.detect_objects_yolov7(processed_frame)

            # Perform object tracking (if previous frame is available)
            if previous_frame is not None:
                tracked_objects = self.track_objects(detections, previous_frame)
            else:
                tracked_objects = []  # No tracking for the first frame

            # Call other analysis functions
            actions = self.analyze_actions(tracked_objects, processed_frame, previous_frame)
            minimap_data = self.analyze_minimap(processed_frame)
            gemini_results = self.analyze_with_gemini(detections)  # Example: Pass detections to Gemini

            # Visualize the results
            visualized_frame = self.draw_visualizations(
                frame.copy(), detections, tracked_objects, actions, minimap_data, gemini_results
            )

            cv2.imshow('CoD: BO6 Analysis', visualized_frame)
            if cv2.waitKey(1) & 0xFF == ord('q'):
                break

            previous_frame = processed_frame  # Update previous frame for tracking

        self.cap.release()
        cv2.destroyAllWindows()

if __name__ == "__main__":
    analyzer = CODAnalyzer("path/to/your/video.mp4")  # Replace with the actual path to your video
    analyzer.run()