requirements.txt
Python 3.9
opencv-python
torch
torchvision
torchaudio
tensorflow
numpy
mediapipe
google-cloud-storage
google-cloud-language
google-cloud-videointelligence
strong-sort
filterpy  # Dependency for StrongSORT
PyQt6